l5cpzt-document-versioning-app-with-history-and-rollback/repository_after/backend/requirements.txt
pytest
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0
python-dotenv>=1.0,<2.0
